Move JQuery Builder types into own utils file
